+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - May be desirable to be used with IDEs that support collapsing code by
     looking at the whitespace of each line.
   - This can be controlled by setting the `GlobalConfig.ASM_INDENTATION` option.
+- Add a way to indentate labels within functions.
+  - May be desirable to be used with IDEs that support collapsing code by
+    looking at the whitespace of each line.
+  - This can be controlled by setting the `GlobalConfig.ASM_INDENTATION_LABELS`
+    option.
 
 ## [1.30.0] - 2024-09-10